feat!: S6b Limits -> RekeyPolicy — default 2^31 packets / 1 TiB, no time trigger
Replaces the unpublished russh::Limits (1 GiB write/read + 1h time
trigger) with RekeyPolicy { max_packets, max_bytes }. The old default
pairing was the 2026-08-10 production volume-rekey accident trigger;
it is now gone. Archival invariants:

1. Public API: russh::Limits deleted (no alias — unpublished, no
   compatibility burden); Config.limits keeps its field name with the
   new type on both server and client. RekeyPolicy::new has no panic
   path; the 1 GiB assert is deleted; thresholds are u64 (1 TiB is
   representable and constructible).
2. Defaults are the I5 hard bounds: max_packets = 2^31,
   max_bytes = 1 TiB. There is no time-based rekey trigger anywhere;
   Encrypted.last_rekey is deleted with it (last_rekey_gen is the
   unrelated supervisor generation counter). In-flight completion
   remains Config.rekey_deadline (S1), untouched.
3. Byte threshold is a single value shared by both directions, but
   accounting and comparison stay per-direction per-epoch
   (out_bytes >= max_bytes and in_bytes >= max_bytes are independent
   predicates over Writer cipher_bytes / Reader bytes_this_epoch) —
   verified not-a-sum by review. S6a gates W1-W6 stay green reading
   the new fields.
4. Behavior-flip gate P3: with Default policy, an injected 2 GiB of
   outbound cipher bytes does NOT enter InKex; the invert (old 1 GiB
   threshold, same injection) must trigger — locking the flip itself.
   Injection lands on the production cipher_bytes atomic via the
   harness override Arc, not an observe slot.
5. Writer cipher_bytes is AtomicU64 end-to-end (no truncating casts);
   harness injection types match.
6. Tests that need volume rekey force it with small explicit
   max_bytes (harness 256 KiB; S0 rekey-stall stays green on that,
   not on defaults); tests that must not rekey use huge thresholds.
   The one test that relied on the 1h time trigger
   (test_kex_shared_secret) now triggers by bytes.
7. zfc constructs Config { ..Default::default() } and never names
   Limits: zero source changes to compile (smoke gate P5). Production
   behavior change to confirm at next zfc integration: no more
   automatic rekey at 1 GiB.
8. Warning baseline stays 21; zero #[allow] added; hooks all
   _test_hooks-gated.

Adversarial review: kimi/k3-256k high r1 GO (review-S6b2-r1-kimi.md).
Pre-audit candidates all excluded with evidence (per-direction
semantics preserved; client flush keeps volume + rekey_wanted
predicates, time-branch removal recorded for S8 alignment; u64
conversion points all widening). Non-blocking observations: stale
rekey_write_limit naming in test_rekey_under_load comments; P4's S0
evidence carried by W6 + P3 invert since S0 forces rekey explicitly.

+++ russh/src/server/writer.rs
@@ -102,7 +102,7 @@ pub struct WriterHooks {
     pub packets_override: Option<Arc<AtomicU64>>,
     /// Share this cipher-bytes atomic with Session (S6a W6 inject).
     #[cfg(feature = "_test_hooks")]
-    pub cipher_bytes_override: Option<Arc<AtomicUsize>>,
+    pub cipher_bytes_override: Option<Arc<AtomicU64>>,
     /// I6/W3: live observation of outbound epoch packet count / seqn.
     #[cfg(feature = "_test_hooks")]
     pub observe: Option<Arc<WriterObserveSlot>>,
@@ -298,7 +298,7 @@ pub struct WriterHandle {
     /// Submitted-but-not-on-socket bytes (queue + out_q + current).
     pending_bytes: Arc<AtomicUsize>,
     /// PacketWriter.buffer().bytes for rekey volume limits.
-    cipher_bytes: Arc<AtomicUsize>,
+    cipher_bytes: Arc<AtomicU64>,
     /// I5 outbound packet count this key-epoch (Session-readable).
     packets_this_epoch: Arc<AtomicU64>,
     capacity: Arc<Notify>,
@@ -621,7 +621,7 @@ impl WriterHandle {
     }
 
     /// Volume counter for rekey write limit (PacketWriter.buffer().bytes).
-    pub fn cipher_bytes(&self) -> usize {
+    pub fn cipher_bytes(&self) -> u64 {
         self.cipher_bytes.load(Ordering::Acquire)
     }
 
@@ -732,9 +732,9 @@ where
     let cipher_bytes = hooks
         .cipher_bytes_override
         .clone()
-        .unwrap_or_else(|| Arc::new(AtomicUsize::new(0)));
+        .unwrap_or_else(|| Arc::new(AtomicU64::new(0)));
     #[cfg(not(feature = "_test_hooks"))]
-    let cipher_bytes = Arc::new(AtomicUsize::new(0));
+    let cipher_bytes = Arc::new(AtomicU64::new(0));
     let cipher_bytes_w = cipher_bytes.clone();
     #[cfg(feature = "_test_hooks")]
     let packets_this_epoch = hooks
@@ -796,7 +796,7 @@ where
                 #[cfg(feature = "_test_hooks")]
                 drain_segs.push_back((pre_len, false));
             }
-            cipher_bytes_w.store(packet_writer.buffer().bytes, Ordering::Release);
+            cipher_bytes_w.store(packet_writer.buffer().bytes as u64, Ordering::Release);
         }
 
         loop {
@@ -1097,7 +1097,7 @@ fn seal_one_payload(
     packet_writer: &mut PacketWriter,
     out_q: &mut VecDeque<Bytes>,
     pending_bytes: &AtomicUsize,
-    cipher_bytes: &AtomicUsize,
+    cipher_bytes: &AtomicU64,
     packets_this_epoch: &AtomicU64,
     p: Bytes,
     hooks: &WriterHooks,
@@ -1205,7 +1205,7 @@ fn seal_one_payload(
         #[cfg(feature = "_test_hooks")]
         drain_segs.push_back((wire_len, as_kex));
     }
-    cipher_bytes.store(packet_writer.buffer().bytes, Ordering::Release);
+    cipher_bytes.store(packet_writer.buffer().bytes as u64, Ordering::Release);
     packets_this_epoch.fetch_add(1, Ordering::AcqRel);
     #[cfg(feature = "_test_hooks")]
     if let Some(ref o) = hooks.observe {
@@ -1216,7 +1216,7 @@ fn seal_one_payload(
 
 fn install_epoch(
     packet_writer: &mut PacketWriter,
-    cipher_bytes: &AtomicUsize,
+    cipher_bytes: &AtomicU64,
     packets_this_epoch: &AtomicU64,
     cipher: Box<dyn SealingKey + Send>,
     outbound_compression: Compression,
@@ -1278,7 +1278,7 @@ fn handle_writer_cmd(
     packet_writer: &mut PacketWriter,
     out_q: &mut VecDeque<Bytes>,
     pending_bytes: &AtomicUsize,
-    cipher_bytes: &AtomicUsize,
+    cipher_bytes: &AtomicU64,
     packets_this_epoch: &AtomicU64,
     evt_tx: &mpsc::UnboundedSender<WriterEvent>,
     shutting_down: &mut bool,
